[Feature] update ep
