+++ fastdeploy/config.py
@@ -274,7 +274,10 @@ class ParallelConfig:
         for key, value in args.items():
             if hasattr(self, key):
                 setattr(self, key, value)
-
+        if isinstance(self.engine_worker_queue_port, str):
+            logger.info(f"engine_worker_queue_port type is str: {self.engine_worker_queue_port}")
+            self.engine_worker_queue_port = [int(port) for port in self.engine_worker_queue_port.split(",")]
+            logger.info(f"engine_worker_queue_port: {self.engine_worker_queue_port}")
         # currently, the expert parallel size is equal data parallel size
         self.expert_parallel_size = self.data_parallel_size
         self.use_ep = self.expert_parallel_size > 1
